requirements.txt
torch
torchvision
transformers
opencv-python
pillow
numpy
pytesseract
pdfplumber
matplotlib
